requirements.txt
# Core
pyspark==3.5.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

# Visualização
matplotlib==3.8.4   

# Config & validação
pyyaml==6.0.1       
pydantic==2.7.1   

# Dev & testes
pytest==8.2.0       

# Jupyter local
jupyterlab==4.2.5
ipykernel==6.29.5
nbconvert==7.16.4
pyppeteer==2.0.0   


# Requisições
requests==2.32.3
tqdm==4.66.4
